requirements.txt
torch
transformers
fastapi
uvicorn
scikit-learn
numpy
sentence-transformers
faiss-cpu
streamlit
shap
matplotlib
